switch from flycam back to the player controller
we might want to use a cylinder collider instead of a capsule. also, its size is a bit strange (we seem to keep sliding? might be too big. either that, or the collider for the chunk is messed up and slopey)
rotating the camera messes up the movement, sadly, still not sure how to fix